+++ app.js
@@ -1,62 +1,48 @@
 const express = require('express')
 const { celebrate, Joi } = require('celebrate');
 const mongoose = require('mongoose');
-const { PORT = 3000 } = process.env;
+
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
 const { errors } = require('celebrate');
-const crypto = require('crypto');
+
 const usersRoute = require('./routes/user')
 const movieRoute = require('./routes/movie')
 const { login, createUser } = require('./controllers/user');
 const { errorLogger, requestLogger } = require('./middlewares/logger');
 const auth = require('./middlewares/auth');
 const NotFoundError = require('./errors/not-found-error');
+const {
+  dataMovies,
+  PORT,
+} = require('./utils/config');
+const errorHandler = require('./middlewares/errorHandler');
+const limiter = require('./middlewares/limiter');
+const {
+  createUserValidate,
+  loginValidate,
+} = require('./middlewares/validation');
+const randomString = require('./middlewares/randomString');
 
-
-mongoose.connect('mongodb://localhost:27017/bitfilmsdb', {
+mongoose.connect(dataMovies, {
   useNewUrlParser: true,
   useUnifiedTopology: true,
 });
 
-
-
 const app = express();
 
-
-
-const randomString = crypto
-  .randomBytes(16) // сгенерируем случайную последовательность 16 байт (128 бит)
-  .toString('hex'); // приведём её к строке
-console.log(randomString);
-
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // за 15 минут
-  max: 100, // можно совершить максимум 100 запросов с одного IP
-});
+// app.use(randomString);
 
 app.use(helmet());
 app.use('/', express.json());
 
 app.use(limiter);
 app.use(requestLogger);
+// eslint-disable-next-line import/no-extraneous-dependencies
 require('dotenv').config();
 
+app.post('/signup', createUserValidate, createUser);
 
-app.post('/signup', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required().min(4),
-    name: Joi.string().min(2).max(30),
-  }),
-}), createUser);
-
-app.post('/signin', celebrate({
-  body: Joi.object().keys({
-    email: Joi.string().required().email(),
-    password: Joi.string().required(),
-  }),
-}), login);
+app.post('/signin', loginValidate, login);
 
 app.use(auth);
 app.use('/', usersRoute);
@@ -66,16 +52,9 @@ app.all('*', (req, res, next) => {
   next(new NotFoundError('ресурс не найден.'));
 });
 
-app.use((err, req, res, next) => {
-  const { statusCode = 500, message } = err;
-  res.status(statusCode).send({ message: statusCode === 500 ? 'На сервере произошла ошибка' : message });
-  next();
-});
+app.use(errorHandler);
 
 app.use(errorLogger);
 app.use(errors());
 
-app.listen(PORT, () => {
-  console.log(`App listening on port ${PORT}`)
-})
-
+app.listen(PORT);
