+++ middlewares/auth.js
@@ -1,7 +1,10 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
 const jwt = require('jsonwebtoken');
 const UnauthorizedError = require('../errors/unauthorized-error');
-
-const { NODE_ENV, JWT_SECRET } = process.env;
+const {
+  NODE_ENV,
+  JWT_SECRET,
+} = require('../utils/config');
 
 module.exports = (req, res, next) => {
   console.log(req.headers);
