+++ routes/user.js
@@ -1,15 +1,9 @@
 const router = require('express').Router();
-const { celebrate, Joi } = require('celebrate');
-
-const {getUserFile, getUserProfileUpdate } =require('../controllers/user');
+const { getUserFile, getUserProfileUpdate } = require('../controllers/user');
+const { userValidate } = require('../middlewares/validation');
 
 router.get('/users/me', getUserFile);
 
-router.patch('/users/me',celebrate({
-  body: Joi.object().keys({
-    name: Joi.string().required().min(2).max(30),
-    email: Joi.string().required().email(),
-  }),
-}), getUserProfileUpdate);
+router.patch('/users/me',userValidate, getUserProfileUpdate);
 
 module.exports = router;
